server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiokafka==0.12.0",
    "fastapi==0.123.10",
    "python-dotenv==1.2.1",
    "uvicorn[standard]==0.38.0",
]
